backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
click==8.1.8
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.6
frozenlist==1.5.0
gotrue==2.11.1
groq==0.13.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
memory-profiler==0.61.0
multidict==6.1.0
packaging==24.2
postgrest==0.19.1
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==13.1
yarl==1.18.3
